refactor: centralize workflow info/badges in list, details pages (#394)
Closes #151
Closes #395
