+++ reana-ui/src/components/LauncherLabel.js
