+++ reana-ui/src/components/WorkflowBadges.module.scss
@@ -8,30 +8,6 @@
   under the terms of the MIT License; see LICENSE file for more details.
 */
 
-.divider {
-  margin: 0 !important;
-  border-color: #f8f8f8 !important;
-}
-
-.badges-and-actions {
-  display: flex;
-  justify-content: space-between;
-  width: 100%;
-  padding: 0.75em 1em;
-
-  .actions {
-    min-width: 22px;
-
-    &:hover {
-      color: #d1d1d1;
-    }
-
-    &.always-visible {
-      visibility: visible;
-    }
-  }
-}
-
 .badgesContainer {
   display: flex;
