+++ reana-ui/src/components/WorkflowProgressCircleBar.js
