+++ reana-ui/src/components/WorkflowProgressCircleBar.module.scss
