+++ reana-ui/src/pages/workflowDetails/WorkflowDetails.module.scss
@@ -1,0 +1,10 @@
+.workflow-details-container {
+  padding-top: 2rem;
+}
+
+.badges-and-actions {
+  display: flex;
+  justify-content: space-between;
+  width: 100%;
+  padding: 0 0.75rem 1rem 1rem;
+}
