--- reana-ui/src/pages/workflowDetails/components/WorkflowInfo.module.scss
@@ -1,68 +1,0 @@
-/*
-  -*- coding: utf-8 -*-
-
-  This file is part of REANA.
-  Copyright (C) 2020, 2022 CERN.
-
-  REANA is free software; you can redistribute it and/or modify it
-  under the terms of the MIT License; see LICENSE file for more details.
-*/
-
-@import "@palette";
-
-.workflow {
-  color: $raven;
-  margin: 2rem 0;
-  font-size: 16px;
-
-  .info {
-    display: flex;
-    justify-content: space-between;
-    padding-bottom: 1em;
-    flex-shrink: 0;
-
-    .details-box {
-      display: flex;
-      align-items: baseline;
-    }
-  }
-
-  .name,
-  .status {
-    font-weight: bold;
-  }
-
-  .name {
-    font-size: 1.3em;
-  }
-
-  .status,
-  .run {
-    font-size: 1.2em;
-  }
-
-  .run {
-    padding: 0 0.8em;
-  }
-
-  .refresh {
-    color: $gray;
-    cursor: pointer;
-    margin-left: 1em;
-    position: relative;
-    top: 10px;
-
-    &:hover {
-      color: $dark-gray;
-    }
-  }
-
-  .notebook {
-    margin-left: 1em;
-    margin: 0.7em;
-  }
-
-  .launcher-label {
-    vertical-align: bottom;
-  }
-}
