+++ reana-ui/src/pages/workflowList/components/WorkflowList.module.scss
@@ -2,7 +2,7 @@
   -*- coding: utf-8 -*-
 
   This file is part of REANA.
-  Copyright (C) 2020, 2021, 2022, 2023 CERN.
+  Copyright (C) 2020, 2021, 2022, 2023, 2024 CERN.
 
   REANA is free software; you can redistribute it and/or modify it
   under the terms of the MIT License; see LICENSE file for more details.
@@ -15,8 +15,33 @@
   justify-content: space-between;
 }
 
+.divider {
+  margin: 0 !important;
+  border-color: #f8f8f8 !important;
+}
+
+.badges-and-actions {
+  display: flex;
+  justify-content: space-between;
+  width: 100%;
+  padding: 0.75em 1em;
+
+  .actions {
+    min-width: 22px;
+    cursor: pointer;
+
+    &:hover {
+      color: #d1d1d1;
+    }
+
+    &.always-visible {
+      visibility: visible;
+    }
+  }
+}
+
 .workflow-details-container {
-  padding: 1em;
+  padding: 1em 1em 0 1em;
   &:hover {
     background-color: lighten($sepia, 3%);
     border-color: darken($sepia, 10%);
